Changed type of forces in ApplySurfaceSampleForces() and ApplyVolumeSampleForces() to float*.

+++ lib/Model/ModelObjects/ModelObject.cxx
@@ -433,11 +433,11 @@ ModelObject
 
 void
 ModelObject
-::ApplySurfaceSampleForces(double* forces) {
+::ApplySurfaceSampleForces(float* forces) {
 }
 
 
 void
 ModelObject
-::ApplyVolumeSampleForces(double* forces) {
+::ApplyVolumeSampleForces(float* forces) {
 }
